+++ cmake/FormatTargets.cmake
@@ -9,6 +9,8 @@ function(pointcloudad_add_format_targets)
     GLOB_RECURSE pointcloudad_format_files
     CONFIGURE_DEPENDS
     "${PROJECT_SOURCE_DIR}/include/*.hpp"
+    "${PROJECT_SOURCE_DIR}/benchmarks/*.cpp"
+    "${PROJECT_SOURCE_DIR}/benchmarks/*.hpp"
     "${PROJECT_SOURCE_DIR}/src/*.cpp"
     "${PROJECT_SOURCE_DIR}/src/*.hpp"
     "${PROJECT_SOURCE_DIR}/tests/*.cpp"
@@ -25,4 +27,3 @@ function(pointcloudad_add_format_targets)
             ${pointcloudad_format_files}
     COMMENT "Checking PointCloudAD source formatting")
 endfunction()
-